+++ src/worker/execution/execution-state.ts
@@ -1,6 +1,8 @@
 export enum ExecutionState {
+    STARTED = 'started',
     INITIALIZING = 'initializing',
     RUNNING = 'running',
     COLLECTING = 'collecting',
+    RESPONSE = 'response',
     DONE = 'done'
 }