+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] // 0.1.1
 
+### Changed
+
+ - using `Worker.terminate()` to cancel running execution immediately
+
 ### Added
 
  - `I18n` for english